[refactor] update import and command list dir
